Updating Homebrew config & dependencies
Trying to be a little more hygienic on my new machine, plus Homebrew now
has a builtin command for setting its required ENV.

The most meaningful change here is to how we think about managing
Postgres. Historically Homebrew's not been a good actor when it comes to
managing Postgres versions, so we're switching to using Postgres.app as
out "version manager".